requirements.txt
# Core utilities
utils-nuuuwan

# Web scraping and HTML parsing
requests
beautifulsoup4

# PDF processing libraries
pypdf
pymupdf
pypdfium2
PyPDF2

# OCR and advanced PDF processing
ocrmypdf

# Data analysis and visualization
matplotlib
scipy

# Development and testing
flake8
pytest
build

google-genai
httpx
dotenv
